requirements.txt
aiogram==3.x
psycopg2-binary
tensorflow>=2.12
natasha
pandas
requests
SpeechRecognition
pydub
